chore: move service/deployments logic to logic/deployments
