+++ content/src/controller/Controller.ts
@@ -8,11 +8,10 @@ import express from 'express'
 import onFinished from 'on-finished'
 import { CURRENT_CATALYST_VERSION, CURRENT_COMMIT_HASH, CURRENT_CONTENT_VERSION } from '../Environment'
 import { getActiveDeploymentsByContentHash } from '../logic/database-queries/deployments-queries'
-import { getContent, isContentAvailable } from '../logic/deployments'
+import { AuditInfo, Deployment, DeploymentOptions, SortingField, SortingOrder } from '../logic/deployment-types'
+import { getContent, getDeployments, isContentAvailable } from '../logic/deployments'
 import { statusResponseFromComponents } from '../logic/status-checks'
 import { toQueryParams } from '../logic/toQueryParams'
-import { getDeployments } from '../service/deployments/deployments'
-import { AuditInfo, Deployment, DeploymentOptions, SortingField, SortingOrder } from '../service/deployments/types'
 import { getPointerChanges } from '../service/pointers/pointers'
 import { PointerChangesFilters } from '../service/pointers/types'
 import {
