+++ content/src/service/pointers/pointers.ts
@@ -1,7 +1,7 @@
 import { PointerChangesSyncDeployment } from '@dcl/schemas'
 import { getHistoricalDeployments, HistoricalDeployment } from '../../logic/database-queries/deployments-queries'
+import { PointerChangesOptions } from '../../logic/deployment-types'
 import { AppComponents } from '../../types'
-import { PointerChangesOptions } from '../deployments/types'
 import { DeploymentPointerChanges } from './types'
 
 const MAX_HISTORY_LIMIT = 500
