+++ content/src/service/pointers/types.ts
@@ -1,5 +1,5 @@
 import { PointerChangesSyncDeployment } from '@dcl/schemas'
-import { DeploymentFilters } from '../../service/deployments/types'
+import { DeploymentFilters } from '../../logic/deployment-types'
 
 export type PointerChangesFilters = Pick<DeploymentFilters, 'from' | 'to' | 'entityTypes'>
