+++ content/test/integration/controller/deployment-pagination.spec.ts
@@ -2,8 +2,8 @@ import { EntityType } from '@dcl/schemas'
 import { fetchJson } from 'dcl-catalyst-commons'
 import { DeploymentField } from '../../../src/controller/Controller'
 import { EnvironmentConfig } from '../../../src/Environment'
+import { DeploymentOptions, SortingField, SortingOrder } from '../../../src/logic/deployment-types'
 import { toQueryParams } from '../../../src/logic/toQueryParams'
-import { DeploymentOptions, SortingField, SortingOrder } from '../../../src/service/deployments/types'
 import { PointerChangesFilters } from '../../../src/service/pointers/types'
 import { makeNoopValidator } from '../../helpers/service/validations/NoOpValidator'
 import { setupTestEnvironment } from '../E2ETestEnvironment'
