+++ content/test/integration/service/deployments/metadata-schema-validations.spec.ts
@@ -1,5 +1,5 @@
 import { EntityType } from '@dcl/schemas'
-import { AuditInfo } from '../../../../src/service/deployments/types'
+import { AuditInfo } from '../../../../src/logic/deployment-types'
 import { DeploymentContext, DeploymentResult } from '../../../../src/service/Service'
 import { AppComponents, EntityVersion } from '../../../../src/types'
 import { makeNoopServerValidator } from '../../../helpers/service/validations/NoOpValidator'
