+++ content/test/integration/service/garbage-collection/garbage-collection.spec.ts
@@ -67,14 +67,14 @@ describe('Integration - Garbage Collection', () => {
     await deployEntitiesCombo(components.deployer, E1)
 
     // Assert all content is available
-    await assertContentIsAvailable(sharedContent, onlyE1Content)
+    await assertContentIsAvailable(components, sharedContent, onlyE1Content)
 
     // Deploy E2
     await deployEntitiesCombo(components.deployer, E2)
 
     // Assert only the shared content is still available
     await awaitUntil(() => assertReportedAsDeletedAre(onlyE1Content))
-    await assertContentIsAvailable(sharedContent)
+    await assertContentIsAvailable(components, sharedContent)
   })
 
   it(`When garbage collection is off, then unused content isn't deleted`, async () => {
@@ -82,7 +82,7 @@ describe('Integration - Garbage Collection', () => {
     await deployEntitiesCombo(components.deployer, E1)
 
     // Assert all content is available
-    await assertContentIsAvailable(sharedContent, onlyE1Content)
+    await assertContentIsAvailable(components, sharedContent, onlyE1Content)
 
     // Deploy E2
     await deployEntitiesCombo(components.deployer, E2)
@@ -91,7 +91,7 @@ describe('Integration - Garbage Collection', () => {
     await sleep(ms('4s'))
 
     // Assert all content is still available
-    await assertContentIsAvailable(sharedContent, onlyE1Content)
+    await assertContentIsAvailable(components, sharedContent, onlyE1Content)
     await assertReportedAsDeletedAre()
   })
 
@@ -104,7 +104,7 @@ describe('Integration - Garbage Collection', () => {
 
     // Assert only the shared content is still available
     await awaitUntil(() => assertReportedAsDeletedAre(onlyE1Content))
-    await assertContentIsAvailable(sharedContent)
+    await assertContentIsAvailable(components, sharedContent)
   })
 
   it(`When entity is not overwritten, then it is not garbage collected`, async () => {
@@ -119,7 +119,7 @@ describe('Integration - Garbage Collection', () => {
 
     // Assert nothing was deleted
     await assertReportedAsDeletedAre()
-    await assertContentIsAvailable(sharedContent, onlyE1Content)
+    await assertContentIsAvailable(components, sharedContent, onlyE1Content)
   })
 
   function assertReportedAsDeletedAre(...fileHashes: string[]) {
@@ -127,8 +127,8 @@ describe('Integration - Garbage Collection', () => {
     return Promise.resolve()
   }
 
-  async function assertContentIsAvailable(...hashes: string[]) {
-    const result = await isContentAvailable(this.components, hashes)
+  async function assertContentIsAvailable(components: AppComponents, ...hashes: string[]) {
+    const result = await isContentAvailable(components, hashes)
     const allAvailable = Array.from(result.values()).every((available) => available)
     assert.ok(allAvailable)
   }
