+++ content/test/unit/logic/deployments-queries.spec.ts
@@ -1,6 +1,6 @@
 import { EntityType } from "@dcl/schemas"
 import { createOrClause, getHistoricalDeploymentsQuery } from '../../../src/logic/database-queries/deployments-queries'
-import { SortingField, SortingOrder } from '../../../src/service/deployments/types'
+import { SortingField, SortingOrder } from '../../../src/logic/deployment-types'
 
 describe('deployments-queries', () => {
   describe('createOrClause', () => {
