+++ content/test/unit/logic/deployments-service.spec.ts
@@ -2,18 +2,16 @@ import { EntityType } from '@dcl/schemas'
 import { createTestMetricsComponent } from '@well-known-components/metrics'
 import { safe } from 'jest-extra-utils'
 import { restore, stub } from 'sinon'
-import { ContentFilesRow } from '../../../../src/logic/database-queries/content-files-queries'
-import { HistoricalDeploymentsRow } from '../../../../src/logic/database-queries/deployments-queries'
-import { MigrationDataRow } from '../../../../src/logic/database-queries/migration-data-queries'
-import { metricsDeclaration } from '../../../../src/metrics'
+import { ContentFilesRow } from '../../../src/logic/database-queries/content-files-queries'
+import { HistoricalDeploymentsRow } from '../../../src/logic/database-queries/deployments-queries'
+import { MigrationDataRow } from '../../../src/logic/database-queries/migration-data-queries'
+import { Deployment, DeploymentOptions, PartialDeploymentHistory } from '../../../src/logic/deployment-types'
 import {
   getCuratedLimit,
-  getCuratedOffset,
-  getDeployments,
-  MAX_HISTORY_LIMIT
-} from '../../../../src/service/deployments/deployments'
-import { Deployment, DeploymentOptions, PartialDeploymentHistory } from '../../../../src/service/deployments/types'
-import { AppComponents } from '../../../../src/types'
+  getCuratedOffset, getDeployments, MAX_HISTORY_LIMIT
+} from '../../../src/logic/deployments'
+import { metricsDeclaration } from '../../../src/metrics'
+import { AppComponents } from '../../../src/types'
 
 describe('deployments service', () => {
   describe('getDeployments', () => {
